+++ ChangeLog.md
@@ -12,6 +12,15 @@
 
 
 
+## future
+
+- **bin**
+  - [ ] 是否增加可选的 *cache* 选项，实现对命令行程序缓存，增加程序响应。
+
+
+
+
+
 ## v1.1.x
 
 - [ ] 内置文本使用 embed 包重写，如拼音等。
@@ -36,7 +45,7 @@
   - +) `CLI.CliAppCompleteInterface`, 并且使其支持系统的内置的固定方法如：`construct > command > help > unmatched > index`（顺序规则）.
   - +) `CliCmd.ArgStringSlice` 用于获取数组式参数
   - +) `Cmd2StringMap` 用户数实现命令字符串转golang风格字符串，并使用其优化命令行路由
-  - +) `CommandAlias/SubCommandAlias` 添加内置的别名机制（适用于 `AppCmd-style`）。
+  - +) `CommandAlias/SubCommandAlias/CommandAliasAll/SubCommandAliasAll` 添加内置的别名机制（适用于 `AppCmd-style`）。
   - -) 移除遗弃标注项：`FormatStr`、`CliCmd.HasSetting`
   - optimize) `bin.FormatTable` 使用**reflect**扩展器支持任意2维度数组格式化。
   - optimze) `CLI.router` 路由重写将过程化的代码转换为多函数调用，未定义方法注册检测更广泛。